Bug half time zone (#1313)
* Fix fractional timezone offsets

* Enable relevant FHIRPath tests

+++ Src/java/cql-to-elm/src/main/java/org/cqframework/cql/cql2elm/Cql2ElmVisitor.java
@@ -976,7 +976,7 @@ public class Cql2ElmVisitor extends CqlPreprocessorElmCommonVisitor {
                         }
 
                         result.setTimezoneOffset(libraryBuilder.createLiteral(
-                                (double) (hourOffset + (minuteOffset / 60)) * offsetPolarity));
+                                (double) (hourOffset + ((double) minuteOffset / 60)) * offsetPolarity));
                     } else {
                         if (matcher.group(26) != null) {
                             int hourOffset = Integer.parseInt(matcher.group(26));
